Added shape criteria code
